refactor: Swiper 코드 스플리팅 및 동적 로딩
