+++ src/app/globals.css
@@ -2,10 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 
-@import "swiper/css";
-@import "swiper/css/pagination";
-@import "swiper/css/navigation";
-
 @layer utilities {
   .shadow-card {
     @apply shadow-[0_4px_20px_rgba(0,0,0,0.08)];
