--- src/components/domains/main/mainContent/MainContentButton/MainContentButtons.tsx
@@ -1,72 +1,0 @@
-"use client";
-
-import { Autoplay, Navigation, Pagination } from "swiper/modules";
-import { Swiper, SwiperSlide } from "swiper/react";
-import { MainContentButton } from "./MainContentButton";
-import { MAIN_BUTTONS } from "@/lib/constants/main";
-
-export const MainContentButtons = () => {
-  return (
-    <>
-      {/* Desktop/Tablet: Grid Layout */}
-      <div className="pc:flex hidden gap-7">
-        {MAIN_BUTTONS.map((button, index) => (
-          <MainContentButton
-            key={index}
-            link={button.link}
-            title={button.title}
-            linkText={button.linkText}
-            bgImage={button.bgImage}
-            isPriority={index === 0}
-          />
-        ))}
-      </div>
-
-      {/* Mobile: Swiper */}
-      <div className="pc:hidden block relative w-full tablet:px-5 px-4 min-h-[140px] tablet:min-h-[200px]">
-        <Swiper
-          spaceBetween={10}
-          slidesPerView={1}
-          modules={[Pagination, Autoplay, Navigation]}
-          pagination={{
-            clickable: true,
-          }}
-          navigation={{
-            nextEl: ".custom-next",
-            prevEl: ".custom-prev",
-          }}
-          autoplay={{ delay: 5000, disableOnInteraction: false }}
-          observer={true}
-          observeParents={true}
-          watchOverflow={false}
-          updateOnWindowResize={true}
-          resistanceRatio={0.85}
-        >
-          {MAIN_BUTTONS.map((button, index) => {
-            const titleLines = button.title.split("\n");
-            return (
-              <SwiperSlide key={index}>
-                <MainContentButton
-                  link={button.link}
-                  title={
-                    <>
-                      {titleLines.map((line, i) => (
-                        <span key={i}>
-                          {line}
-                          {i < titleLines.length - 1 && <br />}
-                        </span>
-                      ))}
-                    </>
-                  }
-                  linkText={button.linkText}
-                  bgImage={button.bgImage}
-                  isPriority={index === 0}
-                />
-              </SwiperSlide>
-            );
-          })}
-        </Swiper>
-      </div>
-    </>
-  );
-};
