+++ src/components/domains/main/mainContent/MainContentButtons/MainContentButtonsSwiper.module.css
@@ -1,3 +1,16 @@
+.swiper-pagination-bullet {
+  width: 4px;
+  height: 4px;
+  background: rgba(255, 255, 255, 0.5);
+  opacity: 1;
+}
+.swiper-pagination-bullet-active {
+  background: #dbc0b0;
+  width: 8px;
+  height: 4px;
+  border-radius: 10px;
+}
+
 /* .custom-swiper-pagination {
   position: absolute;
   top: 16px;
@@ -19,16 +32,3 @@
   height: 8px;
   border-radius: 10px;
 } */
-
-.swiper-pagination-bullet {
-  width: 4px;
-  height: 4px;
-  background: rgba(255, 255, 255, 0.5);
-  opacity: 1;
-}
-.swiper-pagination-bullet-active {
-  background: #dbc0b0;
-  width: 8px;
-  height: 4px;
-  border-radius: 10px;
-}
