+++ src/components/domains/main/mainContent/MainContentButtons/index.ts
@@ -1,0 +1,1 @@
+export { MainContentButtons } from "./MainContentButtons";
